+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -36683,7 +36683,8 @@ $EMT_SelectCompleters = {
 }
 
 $EMT_SelectMap = @{
-    "Select"=@("Add-EMTLogsForPlaybackConfiguration",
+    "Select"=@("Add-EMTLogsForChannel",
+               "Add-EMTLogsForPlaybackConfiguration",
                "New-EMTChannel",
                "New-EMTLiveSource",
                "New-EMTPrefetchSchedule",